requirements.txt
# API and HTTP requests
requests==2.31.0
aiohttp==3.9.3

# Data processing and analysis
pandas==2.2.0

# Environment variables
python-dotenv==1.0.1

# Web application
streamlit==1.32.0

# Type hints
typing-extensions==4.9.0

# PDF processing
PyPDF2==3.0.1

# Asyncio
asyncio==3.4.3 